Add support for Georgian Extended Unicode block to all Georgian OTF fonts
Bug: 111550750

Script affected: Georgian (otf files for Noto Sans and Serif)

Changing the cmaps of our all Georgian fonts to map from the new
uppercase letters to the same glyphs as the old lowercase ones. This is
a temporary solution. The future solution needs to provide correct the
uppercase versions of their Mkhedruli counterparts.

No changes in fonts.xml are needed.

The net size increase is 80 bytes:
* modified 5 OTF files (16 bytes added to each)

Test: atest cts/tests/tests/text/src/android/text/cts/FontCoverageTest.java
Test: All tests described below are manual.

There are two categories of tests:
1. Testing to make sure that the existing Georgian texts render the same
   with the new set of fonts
2. Testing to make sure that the new characters are rendered
   exactly like the old lowercase characters.

Files in
https://github.com/googlei18n/noto-source/tree/master/test/Georgian
were used for the testing.

Also
https://ka.wikipedia.org/wiki/%E1%83%A5%E1%83%90%E1%83%A0%E1%83%97%E1%83%A3%E1%83%9A%E1%83%98_%E1%83%94%E1%83%9C%E1%83%90
https://unicode.org/udhr/d/udhr_kat.html

Change-Id: Ic69b1edae9a60b82b450872923fe68902aa24f46